backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn==0.24.0

# Database
sqlalchemy==2.0.32
aiosqlite==0.19.0

# Authentication and Security
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6

# Environment and Configuration
python-dotenv==1.0.0
pydantic-settings==2.4.0

# HTTP Client
httpx==0.25.2

# Validation and Serialization
pydantic==2.9.2
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2
bcrypt==4.1.3

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Monitoring and Logging
structlog==23.2.0
redis>=4.5.0
bleach>=6.0.0
aiofiles>=23.0.0

# Migrations
alembic==1.13.2

# Database Drivers
asyncpg==0.30.0
psycopg[binary]==3.2.10